requirements.txt
-i https://pypi.org/simple
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
croniter==1.3.4
dask==2022.5.0
distributed==2022.5.0
docker==5.0.3
fsspec==2022.3.0
greenlet==1.1.2
heapdict==1.0.1
idna==3.3
importlib-resources==5.7.1
jinja2==3.1.2
joblib==1.1.0
locket==1.0.0
markupsafe==2.1.1
marshmallow==3.15.0
marshmallow-oneofschema==3.0.1
msgpack==1.0.3
mypy-extensions==0.4.3
nltk==3.7
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
partd==1.2.0
pendulum==2.1.2
praw==7.6.0
prawcore==2.3.0
prefect==1.2.1
psutil==5.9.0
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-box==6.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.1
pytzdata==2020.1
pyyaml==6.0
regex==2022.4.24
requests==2.27.1
six==1.16.0
sortedcontainers==2.4.0
sqlalchemy==1.4.36
tabulate==0.8.9
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
update-checker==0.18.0
urllib3==1.26.9
websocket-client==1.3.2
zict==2.2.0
